+++ api/RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # dxApi
 
+## Unreleased
+
+* Revert preferring v3 instance types over v2, until v3 are more widely available.
+
 ## 0.13.15 (2025-11-14)
 
 * Prefer v3 instance types over v2